add entities harcodeo

+++ src/main/java/com/utn/ElBuenSabor/entities/DetalleArticuloManufacturado.java
@@ -5,7 +5,6 @@ import jakarta.persistence.*;
 import lombok.*;
 import jakarta.validation.constraints.NotNull;
 
-import java.math.BigDecimal;
 
 
 @Entity
@@ -19,7 +18,7 @@ public class DetalleArticuloManufacturado extends Base {
 
     @NotNull
     @Column(name = "cantidad", precision = 10, scale = 2)
-    private BigDecimal cantidad;
+    private Double cantidad;
 
     @NotNull
     @ManyToOne()
