+++ src/main/java/com/utn/ElBuenSabor/entities/DetallePedido.java
@@ -5,7 +5,6 @@ import jakarta.persistence.*;
 import lombok.*;
 import jakarta.validation.constraints.NotNull;
 
-import java.math.BigDecimal;
 @Entity
 @Table(name = "detalle_pedido")
 @NoArgsConstructor
@@ -20,11 +19,11 @@ public class DetallePedido extends Base {
 
     @NotNull
     @Column(name = "subtotal", precision = 10, scale = 2)
-    private BigDecimal subtotal;
+    private Double subtotal;
 
     @NotNull
     @Column(name = "subtotal_costo", precision = 10, scale = 2)
-    private BigDecimal subtotalCosto;
+    private Double subtotalCosto;
 
     @ManyToOne()
     @JoinColumn(name = "id_articulo_manufacturado")
